codegen: Do big fix of the monading typing
Before we were using the PGOCaml module, i.e. the default API with the
classical threads. The interface is there defined with “with 'a t =
'a” so OCaml absorbs many typing errors on the monads.
This patch adds an optional parameter for the code generation
 ?(functorize=true)
in order to force the use the PGOCaml_generic.PGOCAML_GENERIC
interface (which treats then monads like real monads).
The whole thing compiles but the patch is still a bit ugly on the
edges.