Simplify Bcl_to_fastq.start and -sys-dry-run
The todo-list scheme which was already over-complex showed its limits
when trying to do a wet run on data-base-related things. Now the
'system' functions are passed as arguments (like in
Assemble_sample_sheet). The corollary is that now the option -wet-run
disappears in favor of a -sys-dry-run option which makes
system-commands 'dry' but not data-base write-accesses.