Make ACL.set_defaults look for administrators
ACL.set_defaults set read-write access to the 'writers' of the
configuration file and to the persons in the database with role
`administrator. The latter are filtered by actual presence on the
machine using the return code of:
  groups <login>

It is a pretty bit commit because now the ACL.Make functor has the
Layout as argument (which also brings many error types) and that
change had to be propagated through the library.

Finally, the module type ACL has been moved from
hitscore_interfaces.mli to hitscore_acl.ml while hitscore_acl.mli
being removed.